+++ packages/cloud-cognitive/src/components/Datagrid/Datagrid/DatagridContent.js
@@ -8,7 +8,7 @@
 import React, { useContext, useEffect, useRef } from 'react';
 import cx from 'classnames';
 import PropTypes from 'prop-types';
-import { DataTable } from '@carbon/react';
+import { TableContainer, Table } from '@carbon/react';
 import { px } from '@carbon/layout';
 import DatagridHead from './DatagridHead';
 import DatagridBody from './DatagridBody';
@@ -16,24 +16,27 @@ import DatagridToolbar from './DatagridToolbar';
 import { handleGridKeyPress } from './addons/InlineEdit/handleGridKeyPress';
 import { carbon, pkg } from '../../../settings';
 import { InlineEditContext } from './addons/InlineEdit/InlineEditContext';
+import { FilterContext, FilterPanel } from './addons/Filtering';
 import { handleGridFocus } from './addons/InlineEdit/handleGridFocus';
 import { useClickOutside } from '../../../global/js/hooks';
 import { useMultipleKeyTracking } from '../../DataSpreadsheet/hooks';
-
-const { TableContainer, Table } = DataTable;
+import { FilterSummary } from '../../FilterSummary';
+import { CLEAR_FILTERS } from './addons/Filtering/constants';
 
 const blockClass = `${pkg.prefix}--datagrid`;
 
 export const DatagridContent = ({ datagridState }) => {
-  const { state, dispatch } = useContext(InlineEditContext);
-  const { activeCellId } = state;
+  const { state: inlineEditState, dispatch } = useContext(InlineEditContext);
+  const { filterTags, EventEmitter, panelOpen } = useContext(FilterContext);
+  const { activeCellId, gridActive, editId } = inlineEditState;
   const {
     getTableProps = () => {},
+    getFilterFlyoutProps,
     withVirtualScroll,
     DatagridPagination,
     isFetching,
     CustomizeColumnsTearsheet,
-    leftPanel,
+    filterProps,
     fullHeightDatagrid,
     verticalAlign = 'center',
     variableRowHeight,
@@ -45,10 +48,10 @@ export const DatagridContent = ({ datagridState }) => {
     DatagridActions,
     totalColumnsWidth,
     gridRef,
+    state,
   } = datagridState;
 
   const rows = (DatagridPagination && datagridState.page) || datagridState.rows;
-  const { gridActive, editId } = state;
   const gridAreaRef = useRef();
   const multiKeyTrackingRef = useRef();
 
@@ -88,14 +91,18 @@ export const DatagridContent = ({ datagridState }) => {
                 handleGridKeyPress({
                   event,
                   dispatch,
-                  state,
+                  inlineEditState,
                   instance: datagridState,
                   keysPressedList,
                   usingMac,
                 })
             : null
         }
-        onFocus={withInlineEdit ? () => handleGridFocus(state, dispatch) : null}
+        onFocus={
+          withInlineEdit
+            ? () => handleGridFocus(inlineEditState, dispatch)
+            : null
+        }
       >
         {!withVirtualScroll ? <DatagridHead {...datagridState} /> : null}
         <DatagridBody {...datagridState} rows={rows} />
@@ -130,6 +137,15 @@ export const DatagridContent = ({ datagridState }) => {
     }
   }, [withInlineEdit, tableId, totalColumnsWidth, datagridState, gridActive]);
 
+  const renderFilterSummary = () =>
+    state.filters.length > 0 && (
+      <FilterSummary
+        className={`${blockClass}__filter-summary`}
+        filters={filterTags}
+        clearFilters={() => EventEmitter.dispatch(CLEAR_FILTERS)}
+      />
+    );
+
   return (
     <>
       <TableContainer
@@ -151,24 +167,35 @@ export const DatagridContent = ({ datagridState }) => {
         description={gridDescription}
       >
         <DatagridToolbar {...datagridState} />
-        <div className={`${blockClass}__table-container`} ref={gridAreaRef}>
-          {leftPanel && leftPanel.isOpen && (
-            <div className={`${blockClass}__datagridLeftPanel`}>
-              {leftPanel.panelContent}
-            </div>
+        <div
+          className={cx(`${blockClass}__table-container`, {
+            [`${blockClass}__table-container--filter-open`]: panelOpen,
+          })}
+          ref={gridAreaRef}
+        >
+          {filterProps?.variation === 'panel' && (
+            <FilterPanel
+              updateMethod="batch"
+              {...getFilterFlyoutProps()}
+              title={filterProps.panelTitle}
+              filterSections={filterProps.sections}
+            />
           )}
-          {withInlineEdit ? (
-            <div ref={multiKeyTrackingRef}>{renderTable()}</div>
-          ) : withVirtualScroll ? (
-            <div
-              className={`${blockClass}__virtualScrollContainer`}
-              ref={gridRef}
-            >
-              {renderTable()}
-            </div>
-          ) : (
-            renderTable()
-          )}
+          <div className={`${blockClass}__table-container-inner`}>
+            {renderFilterSummary()}
+            {withInlineEdit ? (
+              <div ref={multiKeyTrackingRef}>{renderTable()}</div>
+            ) : withVirtualScroll ? (
+              <div
+                className={`${blockClass}__virtualScrollContainer`}
+                ref={gridRef}
+              >
+                {renderTable()}
+              </div>
+            ) : (
+              renderTable()
+            )}
+          </div>
         </div>
       </TableContainer>
       {rows?.length > 0 && !isFetching && DatagridPagination && (
@@ -184,6 +211,7 @@ export const DatagridContent = ({ datagridState }) => {
 DatagridContent.propTypes = {
   datagridState: PropTypes.shape({
     getTableProps: PropTypes.func,
+    getFilterFlyoutProps: PropTypes.func,
     withVirtualScroll: PropTypes.bool,
     DatagridActions: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
     DatagridPagination: PropTypes.oneOfType([
@@ -195,8 +223,8 @@ DatagridContent.propTypes = {
       PropTypes.func,
     ]),
     isFetching: PropTypes.bool,
-    leftPanel: PropTypes.object,
     fullHeightDatagrid: PropTypes.bool,
+    filterProps: PropTypes.object,
     variableRowHeight: PropTypes.bool,
     useDenseHeader: PropTypes.bool,
     withInlineEdit: PropTypes.bool,
@@ -208,5 +236,8 @@ DatagridContent.propTypes = {
     tableId: PropTypes.string,
     totalColumnsWidth: PropTypes.number,
     gridRef: PropTypes.object,
+    setAllFilters: PropTypes.func,
+    getFilterProps: PropTypes.func,
+    state: PropTypes.object,
   }),
 };