+++ packages/cloud-cognitive/src/components/Datagrid/Datagrid/DatagridToolbar.js
@@ -5,20 +5,19 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import React, { useEffect, useState, useContext } from 'react';
+import React, { useEffect, useState } from 'react';
 import { Add, OverflowMenuVertical } from '@carbon/react/icons';
-import { DataTable, TableBatchActions, TableBatchAction } from '@carbon/react';
+import {
+  TableToolbar,
+  TableBatchActions,
+  TableBatchAction,
+} from '@carbon/react';
 import { useResizeDetector } from 'react-resize-detector';
 import { ButtonMenu, ButtonMenuItem } from '../../ButtonMenu';
 import { pkg, carbon } from '../../../settings';
 import cx from 'classnames';
-import { FilterSummary } from '../../FilterSummary';
-import { FilterContext } from './addons/Filtering/FilterProvider';
-import { CLEAR_FILTERS } from './addons/Filtering/constants';
 
 const blockClass = `${pkg.prefix}--datagrid`;
-
-const { TableToolbar } = DataTable;
 
 const DatagridBatchActionsToolbar = (datagridState, width, ref) => {
   const [displayAllInMenu, setDisplayAllInMenu] = useState(false);
@@ -157,26 +156,24 @@ const DatagridBatchActionsToolbar = (datagridState, width, ref) => {
 
 const DatagridToolbar = (datagridState) => {
   const { width, ref } = useResizeDetector();
-  const { DatagridActions, DatagridBatchActions, batchActions, state } =
+  const { DatagridActions, DatagridBatchActions, batchActions, rowSize } =
     datagridState;
-  const { filterTags, EventEmitter } = useContext(FilterContext);
 
-  const renderFilterSummary = () =>
-    state.filters.length > 0 && (
-      <FilterSummary
-        filters={filterTags}
-        clearFilters={() => EventEmitter.dispatch(CLEAR_FILTERS)}
-      />
-    );
+  const getRowHeight = rowSize ? rowSize : 'lg';
 
   return batchActions && DatagridActions ? (
-    <div ref={ref} className={`${blockClass}__table-toolbar`}>
+    <div
+      ref={ref}
+      className={cx(
+        `${blockClass}__table-toolbar`,
+        `${blockClass}__table-toolbar--${getRowHeight}`
+      )}
+    >
       <TableToolbar>
         {DatagridActions && DatagridActions(datagridState)}
         {DatagridBatchActionsToolbar &&
           DatagridBatchActionsToolbar(datagridState, width, ref)}
       </TableToolbar>
-      {renderFilterSummary()}
     </div>
   ) : DatagridActions ? (
     <div className={`${blockClass}__table-toolbar`}>
@@ -184,7 +181,6 @@ const DatagridToolbar = (datagridState) => {
         {DatagridActions && DatagridActions(datagridState)}
         {DatagridBatchActions && DatagridBatchActions(datagridState)}
       </TableToolbar>
-      {renderFilterSummary()}
     </div>
   ) : null;
 };